requirements.txt
asgiref==3.5.2
black==22.10.0
Django==4.1.2
django-cors-headers==3.13.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
gunicorn==20.1.0
Pillow==9.3.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
python-decouple==3.6
sqlparse==0.4.3
